requirements.txt
line-bot-sdk>=3.8.0
flask>=3.0.2
requests>=2.28.0
